Getting started/requirements.txt
# Standard
python-dotenv

# Redis
redis